feat: implemented LiquidChat as module instead
- added JWT support
- improved messages and error handling
- added more events
