+++ src/main/kotlin/net/ccbluex/liquidbounce/event/EventManager.kt
@@ -85,8 +85,10 @@ val ALL_EVENT_CLASSES: Array<KClass<out Event>> = arrayOf(
     ValueChangedEvent::class,
     ToggleModuleEvent::class,
     NotificationEvent::class,
+    ClientChatStateChange::class,
     ClientChatMessageEvent::class,
     ClientChatErrorEvent::class,
+    ClientChatJwtTokenEvent::class,
     WorldChangeEvent::class,
     AltManagerUpdateEvent::class,
     VirtualScreenEvent::class,
