+++ src/main/resources/assets/liquidbounce/lang/en_us.json
@@ -276,11 +276,11 @@
   "liquidbounce.generic.locked": "The module is locked and cannot be enabled. It is likely not bypassing.",
   "liquidbounce.liquidchat.authenticationFailed": "Authentication failed! LiquidChat requires Minecraft premium account.",
   "liquidbounce.liquidchat.jwtTokenReceived": "Received JWT token! Relogging...",
+  "liquidbounce.liquidchat.notConnected": "Not connected to the chat server! Turn on the LiquidChat module!",
+  "liquidbounce.liquidchat.notLoggedIn": "Not authenticated! You can only chat with a premium Minecraft account, sorry!",
   "liquidbounce.liquidchat.states.connected": "Connected to chat server!",
-  "liquidbounce.liquidchat.states.connecting": "Connecting to chat server...",
   "liquidbounce.liquidchat.states.disconnected": "Disconnected from chat server!",
   "liquidbounce.liquidchat.states.loggedIn": "Authenticated!",
-  "liquidbounce.liquidchat.states.loggingIn": "Authenticating...",
   "liquidbounce.liquidchat.writeMessage": "To send a message: %s",
   "liquidbounce.liquidchat.writePrivateMessage": "To send a private message: %s",
   "liquidbounce.module.ESP.description": "Allows you to see targets through walls.",
